Extract residual persistence and restore helpers
